Create SNP_to_rsIDs
This is a script that helps to convert your variants (SNPs) to respective rsIDs. This particular script is adapted to hg17 Homo Sapiens version of genome. Make sure that you have Chr and BP in your sample dataset. The examples of sample dataset is in the file name sample_dataset.txt